Fix description based on comment
"description": "This is a plugin that fetches contacts from iCloud and manages a file for each contact",
Avoid including sentences like "This is a plugin that does ...." in your description.
We have published a guide for plugin descriptions to follow here: https://docs.obsidian.md/Plugins/Releasing/Submission+requirements+for+plugins#Keep+plugin+descriptions+short+and+simple.